Prevent endless loop of disabled upstream notifications

+++ database/migrations/2024_01_10_155555_add_bank_transaction_nordigen_field.php
@@ -1,12 +1,13 @@
 <?php
 
+use App\Utils\Ninja;
 use App\Models\Account;
 use App\Models\BankIntegration;
 use App\Models\BankTransaction;
-use App\Repositories\BankTransactionRepository;
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+use App\Repositories\BankTransactionRepository;
 
 return new class extends Migration {
     /**
@@ -22,19 +23,22 @@ return new class extends Migration {
             $table->text('nordigen_transaction_id')->nullable();
         });
 
-        // remove invalid transactions
-        BankIntegration::query()->where('integration_type', BankIntegration::INTEGRATION_TYPE_NORDIGEN)->cursor()->each(function ($bank_integration) {
-            $bank_integration->from_date = now()->subDays(90);
-            $bank_integration->save();
+        if(Ninja::isSelfHost())
+        {
+            // remove invalid transactions
+            BankIntegration::query()->where('integration_type', BankIntegration::INTEGRATION_TYPE_NORDIGEN)->cursor()->each(function ($bank_integration) {
+                $bank_integration->from_date = now()->subDays(90);
+                $bank_integration->save();
 
-            BankTransaction::query()->where('bank_integration_id', $bank_integration->id)->cursor()->each(function ($bank_transaction) {
-                if ($bank_transaction->invoiceIds != '' || $bank_transaction->expense_id != '')
-                    return;
+                BankTransaction::query()->where('bank_integration_id', $bank_integration->id)->cursor()->each(function ($bank_transaction) {
+                    if ($bank_transaction->invoiceIds != '' || $bank_transaction->expense_id != '')
+                        return;
 
-                $btrepo = new BankTransactionRepository();
-                $btrepo->delete($bank_transaction);
+                    $btrepo = new BankTransactionRepository();
+                    $btrepo->delete($bank_transaction);
+                });
             });
-        });
+        }
     }
 
     /**
